+++ Back-end/src/main/java/com/simplon/dvdStore/repositories/ClientRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * The interface Client repository.
  */
-public interface ClientRepository extends CrudRepository<ClientRepositoryModelSQL, Long> {
+public interface ClientRepository extends CrudRepository<ClientRepositoryModelSQL, Integer> {
 }
